# openapi.yaml
openapi: 3.0.3
info:
  title: freeswitch-security REST API
  version: 1.0.0
  description: |
    Control and monitoring API for **freeswitch-security**, the FreeSWITCH ESL →
    ipset IP blocker for the VoiceTel fleet.

    ## Role in the fleet
    This daemon enforces bans locally (kernel ipset behind one iptables match-set
    DROP rule) and participates in the central **chanDaemon** ban repository (D39):
    every enacted ban is reported to chanDaemon, and chanDaemon pushes
    customer/operator unbans back to `DELETE /api/v1/ips/{ip}/block`. See
    `chandaemon.md`.

    ## Authentication / access control
    There is no token. **State-changing** endpoints (`POST`/`PUT`/`PATCH`/`DELETE`)
    are restricted by a **source-IP allow-list**
    (`security.chandaemon.allowed_api_ips`, IPs and CIDRs) enforced in Go against
    the raw connection address (never `X-Forwarded-For`). It defaults to the
    chanDaemon nodes plus loopback, so the API is restricted by default; an empty
    list leaves it unrestricted. Read-only `GET` endpoints are always open; a
    non-allowed source receives `403`. For production, bind to localhost or front
    it with an authenticating reverse proxy.

    Request and response property names are `camelCase`, consistent with the
    rest of the fleet (chanDaemon, opensips-journal-blocker).
  license:
    name: MIT
    url: https://opensource.org/license/mit
  contact:
    name: VoiceTel Platform Engineering
    url: https://voicetel.com
servers:
...
paths:
...
  /cache/security/clear:
    post:
      tags: [Cache]
      summary: Clear the security response cache
      description: Drops the cached `/security/stats` response. State-changing, so gated by the source-IP allow-list.
      operationId: clearSecurityCache
      responses:
        "200":
          description: Cache cleared.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Health" }
              example: { status: "security cache cleared" }
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/ips/{ip}/block:
    delete:
      tags: [chanDaemon]
      summary: Lift a ban (chanDaemon unban fan-out)
      description: |
        Removes the IP from the blacklist and the kernel ipset. This is the
        receiver for chanDaemon's unban fan-out: chanDaemon sends an
        unauthenticated `DELETE` here when a customer/operator lifts a ban, so
        the only gate is the source-IP allow-list. The operation is idempotent —
        lifting an IP that is not banned still returns `200`.
      operationId: unblockIP
      parameters:
        - $ref: "#/components/parameters/IPPath"
      responses:
        "200":
          description: Ban lifted (or the IP was not banned).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnbanResponse" }
              example: { ip: "203.0.113.50", banned: false, message: "IP unblocked successfully" }
        "400":
          $ref: "#/components/responses/InvalidIP"
        "403":
          $ref: "#/components/responses/Forbidden"
  /security/whitelist:
    get:
      tags: [Whitelist]
      summary: List whitelisted IPs
      description: Returns the active whitelist entries, keyed by IP address.
      operationId: listWhitelist
      responses:
        "200":
          description: Whitelisted IPs keyed by address.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WhitelistMap" }
              example:
                "203.0.113.7":
                  ip: "203.0.113.7"
                  userId: "2017301000"
                  domain: "x.example"
                  permanent: false
    post:
      tags: [Whitelist]
      summary: Add an IP to the whitelist
      description: Adds an IP to the whitelist; rejects a malformed IP with 400.
      operationId: addWhitelist
      requestBody:
        required: true
        description: The IP to whitelist and its optional user/domain/permanent flags.
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WhitelistRequest" }
            example: { ip: "203.0.113.7", userId: "2017301000", domain: "x.example", permanent: false }
      responses:
        "200":
          description: IP whitelisted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatusMessage" }
              example: { status: "success", message: "IP 203.0.113.7 added to whitelist for 2017301000@x.example" }
        "400":
          $ref: "#/components/responses/InvalidIP"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /security/whitelist/{ip}:
    get:
      tags: [Whitelist]
      summary: Check whitelist status of an IP
      description: Reports whether the IP is whitelisted, with the associated user/domain when it is.
      operationId: getWhitelistEntry
      parameters:
        - $ref: "#/components/parameters/IPPath"
      responses:
        "200":
          description: Whitelist status.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WhitelistCheck" }
              example: { ip: "203.0.113.7", whitelisted: true, userId: "2017301000", domain: "x.example" }
    delete:
      tags: [Whitelist]
      summary: Remove an IP from the whitelist
      description: Removes the given IP from the whitelist; a no-op if it is absent.
      operationId: removeWhitelist
      parameters:
        - $ref: "#/components/parameters/IPPath"
      responses:
        "200":
          description: IP removed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatusMessage" }
              example: { status: "success", message: "IP 203.0.113.7 removed from whitelist" }
        "403":
          $ref: "#/components/responses/Forbidden"
  /security/blacklist:
    get:
      tags: [Blacklist]
      summary: List blacklisted IPs
      description: Returns the active blacklist entries (bans), keyed by IP address.
      operationId: listBlacklist
      responses:
        "200":
          description: Blacklisted IPs keyed by address.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlacklistMap" }
              example:
                "203.0.113.50":
                  ip: "203.0.113.50"
                  reason: "Exceeded max failed registrations (5)"
                  permanent: false
    post:
      tags: [Blacklist]
      summary: Manually blacklist an IP
      description: |
        Bans the IP (added to the blacklist and the kernel ipset). When
        chanDaemon reporting is enabled the ban is also reported. Rejected if the
        IP is whitelisted or in a trusted network.
      operationId: addBlacklist
      requestBody:
        required: true
        description: The IP to ban, its reason, and an optional permanent flag.
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BlacklistRequest" }
            example: { ip: "203.0.113.50", reason: "manual block", permanent: false }
      responses:
        "200":
          description: IP blacklisted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatusMessage" }
              example: { status: "success", message: "IP 203.0.113.50 added to blacklist" }
        "400":
          description: Invalid IP, or the IP is whitelisted / in a trusted network.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "cannot blacklist whitelisted IP: 203.0.113.50" }
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /security/iptables:
    get:
      tags: [Tracking]
      summary: Firewall (ipset) view
      description: The IPs currently in the kernel ipset and the chain holding the match-set DROP rule.
      operationId: getIPTablesInfo
      responses:
        "200":
          description: Firewall state.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IPTablesInfo" }
              example: { chain: "INPUT", ipset: "freeswitch-security", blockedIps: ["203.0.113.50"], count: 1 }
        "500":
          $ref: "#/components/responses/ServerError"
...
  /security/esl/log-level:
    post:
      tags: [ESL]
      summary: Set the ESL log level
      description: Sets the runtime log level (error, info, debug, trace).
      operationId: setESLLogLevel
      requestBody:
        required: true
        description: The log level to apply.
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LogLevelRequest" }
            example: { level: "debug" }
      responses:
        "200":
          description: Log level set.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatusMessage" }
              example: { status: "success", message: "ESL log level set to debug" }
        "400":
          description: Missing the required `level` field.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "Key: 'level' Error:Field validation for 'level' failed on the 'required' tag" }
        "403":
          $ref: "#/components/responses/Forbidden"
  /security/esl/reconnect:
    post:
      tags: [ESL]
      summary: Force an ESL reconnection
      description: Initiates a reconnection to the FreeSWITCH event socket.
      operationId: reconnectESL
      responses:
        "200":
          description: Reconnection initiated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatusMessage" }
              example: { status: "success", message: "ESL reconnection initiated" }
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /security/untrusted-networks:
    get:
      tags: [Untrusted Networks]
      summary: List untrusted-domain patterns
      description: Returns every exact-match domain pattern that triggers a ban on a failed registration.
      operationId: listUntrustedNetworks
      responses:
        "200":
          description: The untrusted-domain patterns.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PatternList" }
              example: ["evil.example", "scanner.example"]
    post:
      tags: [Untrusted Networks]
      summary: Add an untrusted-domain pattern
      description: Adds an exact-match domain pattern to the untrusted list.
      operationId: addUntrustedNetwork
      requestBody:
        required: true
        description: The pattern to add.
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatternRequest" }
            example: { pattern: "evil.example" }
      responses:
        "200":
          description: Pattern added.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatusMessage" }
              example: { status: "success", message: "Pattern 'evil.example' added to untrusted networks" }
        "400":
          description: Missing pattern, or the pattern already exists.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "untrusted pattern already exists: \"evil.example\"" }
        "403":
          $ref: "#/components/responses/Forbidden"
  /security/untrusted-networks/{pattern}:
    delete:
      tags: [Untrusted Networks]
      summary: Remove an untrusted-domain pattern
      description: Removes a pattern from the untrusted list.
      operationId: removeUntrustedNetwork
      parameters:
        - name: pattern
          in: path
          required: true
          description: The exact pattern to remove.
          schema: { type: string, example: "evil.example" }
      responses:
        "200":
          description: Pattern removed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatusMessage" }
              example: { status: "success", message: "Pattern 'evil.example' removed from untrusted networks" }
        "400":
          description: The pattern was not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "untrusted pattern not found: \"evil.example\"" }
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  parameters:
    IPPath:
      name: ip
      in: path
      required: true
      description: IPv4 address.
      schema: { type: string, format: ipv4, example: "203.0.113.50" }
  responses:
    InvalidIP:
      description: Malformed IP address.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "invalid IP address: not-an-ip" }
    Forbidden:
      description: The source IP is not in the API allow-list.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "source IP not permitted for this operation" }
...
  schemas:
    Health:
      type: object
      description: A single status word, with the build version on /health.
      properties:
        status: { type: string, example: "ok", description: "status word" }
        version: { type: string, example: "v1.5.3", description: "build version (/health only; \"dev\" for an un-stamped build)" }
    Error:
      type: object
      description: Plain error envelope returned on a non-2xx response.
      properties:
        error: { type: string, example: "source IP not permitted for this operation", description: "human-readable error detail" }
    StatusMessage:
      type: object
      description: Acknowledgement of a successful mutation.
      properties:
        status: { type: string, example: "success", description: "always \"success\" on a 200" }
        message: { type: string, example: "IP 203.0.113.50 added to blacklist", description: "human-readable confirmation" }
    UnbanResponse:
      type: object
      description: Result of lifting a ban via the chanDaemon unban endpoint.
      properties:
        ip: { type: string, example: "203.0.113.50", description: "the IP that was unblocked" }
        banned: { type: boolean, example: false, description: "always false on success" }
        message: { type: string, example: "IP unblocked successfully", description: "unblock confirmation" }
...
    WhitelistRequest:
      type: object
      description: A request to add an IP to the whitelist.
      required: [ip]
      properties:
        ip: { type: string, format: ipv4, example: "203.0.113.7", description: "the IP to whitelist" }
        userId: { type: string, example: "2017301000", description: "associated SIP user (optional)" }
        domain: { type: string, example: "x.example", description: "associated domain (optional; defaults to freeswitch.default_domain)" }
        permanent: { type: boolean, example: false, description: "never expire this entry" }
...
    WhitelistCheck:
      type: object
      description: Whitelist status of one IP; user/domain present only when whitelisted.
      properties:
        ip: { type: string, example: "203.0.113.7", description: "the IP queried" }
        whitelisted: { type: boolean, example: true, description: "true if currently whitelisted" }
        userId: { type: string, example: "2017301000", description: "associated SIP user (present when whitelisted)" }
        domain: { type: string, example: "x.example", description: "associated domain (present when whitelisted)" }
...
    PatternList:
      type: array
      description: Untrusted-domain patterns.
      items: { type: string, example: "evil.example" }
    WhitelistMap:
      type: object
      description: Active whitelist entries keyed by IP address.
      additionalProperties:
        $ref: "#/components/schemas/WhitelistEntry"
    BlacklistMap:
      type: object
      description: Active blacklist entries keyed by IP address.
      additionalProperties:
        $ref: "#/components/schemas/BlacklistEntry"
...